New option to use a single, fixed, 1GB huge page for host/coprocessor interface buffers
This feature is driven by the new CHtHifParams element,
m_hostMemHugePageBase.  This parameter is null by default.  However,
when set, it is used as the addr argument to the mmap() function.
In this implementation, mmap() is called with the following flags:
MAP_ANONYMOUS | MAP_POPULATE | MAP_HUGETLB | MAP_FIXED.
In this implementation, it is assumed that the huge page size has been
set to 1GB.
The benefit of using this feature is that performance is improved (at
the cost of almost 1GB of host memory) for the application that suffers
frequent tlb pid invalidates.